requirements/prod.txt
boto3==1.34.33
Django==5.0.1
django-allauth==0.60.1
django-storages==1.14.2
pillow==10.2.0
psycopg2==2.9.9
python-dotenv==1.0.1
